requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
click==8.1.8; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
importlib-metadata==8.5.0; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
packaging==25.0; python_version >= '3.8'
werkzeug==3.0.6; python_version >= '3.8'
zipp==3.20.2; python_version >= '3.8'
mailjet-rest==1.3.4
python-dotenv==1.0.0
